backend/src/requirements/base.txt
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
pillow==10.3.0
python-dotenv==1.0.1
black==24.4.2
djangorestframework==3.15.2
dj-rest-auth==6.0.0
django-allauth[socialaccount]==64.2.0
djangorestframework-simplejwt==5.3.1
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0